Update bulk_export_monitors.py
Monitor name is now a mandatory parameter in the yaml definition when using "montecarlo apply" in MaC. I have added logic to optionally export the monitor name. This is because, by default, the exportMonteCarloConfigTemplates API doesn't return name unless an "exportName" argument (bool) is also included. This version asks the client to choose whether to export it or not (y/N).